+++ src/main/java/com/github/lipiridi/spotless/applier/onSave/SpotlessOnSaveActionInfoBase.java
@@ -1,7 +1,7 @@
 // Copyright 2000-2021 JetBrains s.r.o. and contributors. Use of this source code is governed by the
 // Apache 2.0 license that can be found in the LICENSE file.
 
-package com.github.lipiridi.spotless.applier.config;
+package com.github.lipiridi.spotless.applier.onSave;
 
 import com.intellij.codeInsight.CodeInsightBundle;
 import com.intellij.ide.actionsOnSave.ActionOnSaveComment;
