+++ src/main/java/com/github/lipiridi/spotless/applier/onSave/SpotlessOnSaveInfoProvider.java
@@ -1,4 +1,4 @@
-package com.github.lipiridi.spotless.applier.config;
+package com.github.lipiridi.spotless.applier.onSave;
 
 import com.intellij.ide.actionsOnSave.ActionOnSaveContext;
 import com.intellij.ide.actionsOnSave.ActionOnSaveInfo;
