+++ src/main/java/com/github/lipiridi/spotless/applier/onSave/SpotlessOnSaveOptions.java
@@ -1,7 +1,7 @@
 // Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the
 // Apache 2.0 license.
 
-package com.github.lipiridi.spotless.applier.config;
+package com.github.lipiridi.spotless.applier.onSave;
 
 import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.components.PersistentStateComponent;
