+++ src/main/resources/META-INF/plugin.xml
@@ -50,7 +50,7 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <actionOnSaveInfoProvider id="SpotlessOnSaveInfoProvider"
-                                  implementation="com.github.lipiridi.spotless.applier.config.SpotlessOnSaveInfoProvider"
+                                  implementation="com.github.lipiridi.spotless.applier.onSave.SpotlessOnSaveInfoProvider"
                                   order="after FormatOnSaveInfoProvider, before BuildOnSaveInfoProvider"/>
         <actionOnSave id="SpotlessActionOnSave"
                       implementation="com.github.lipiridi.spotless.applier.actions.ReformatActionOnSave"
@@ -58,6 +58,6 @@
     </extensions>
 
     <extensionPoints>
-        <extensionPoint name="formatOnSaveOptions.spotlessProvider" interface="com.github.lipiridi.spotless.applier.config.SpotlessOnSaveOptionsBase$DefaultsProvider" dynamic="true"/>
+        <extensionPoint name="formatOnSaveOptions.spotlessProvider" interface="com.github.lipiridi.spotless.applier.onSave.SpotlessOnSaveOptionsBase$DefaultsProvider" dynamic="true"/>
     </extensionPoints>
 </idea-plugin>
